third_party/WebKit/LayoutTests/imported/wpt/html/semantics/forms/the-datalist-element/datalistoptions-expected.txt

This is a testharness.js-based test.
FAIL options label/value assert_equals: expected 5 but got 3
PASS If an element has a datalist element ancestor, it is barred from constraint validation 
Harness: the test ran to completion.
